+++ time4j-android/src/main/java/net/time4j/calendar/HijriCalendar.java
@@ -386,6 +386,7 @@ public final class HijriCalendar
      * <p>The supported range of islamic years is 1-1600. </p>
      *
      * @since   3.6/4.4
+     * @deprecated  scheduled for future removal, as workaround, users can use ICU4J directly in Android
      */
     /*[deutsch]
      * Der Name der astronomischen ICU4J-Variante.
@@ -393,7 +394,9 @@ public final class HijriCalendar
      * <p>Der unterst&uuml;tze Wertebereich der islamischen Jahre ist 1-1600. </p>
      *
      * @since   3.6/4.4
+     * @deprecated  scheduled for future removal, as workaround, users can use ICU4J directly in Android
      */
+    @Deprecated
     public static final String VARIANT_ICU4J = "islamic-icu4j";
 
     /**